+++ PlantUmlViewer/PlantUmlViewer/Forms/PreviewWindow.cs
@@ -154,6 +154,23 @@ namespace PlantUmlViewer.Forms
             ImageBox_ZoomChanged(this, null);
         }
 
+        protected override void OnHandleCreated(EventArgs e)
+        {
+            /*
+             * Notepad++ may freezes if the plugins control is docked out and loosing focus
+             *  - https://github.com/notepad-plus-plus/notepad-plus-plus/issues/8723
+             *  - https://forums.codeguru.com/showthread.php?412421-endless-messages-WM_GETDLGCODE-when-focus-is-lost
+             *  - http://robs-got-a-blog.blogspot.com/2011/09/net-custom-control-locks-up-when-losing.html
+             *  - https://social.msdn.microsoft.com/Forums/en-US/dca98588-039a-44f8-b15c-86fd9ac415c6/how-can-i-resolve-datagridview-and-cdialog-bug?forum=winformsdatacontrols
+             *  Fix: Set WS_EX_CONTROLPARENT to elemens which may are in focus
+             */
+            foreach (Button button in this.GetChildren<Button>(10))
+            {
+                NativeWindowApi.ModifyStyleEx(button.Handle, 0, NativeWindowApi.WS_EX_CONTROLPARENT);
+            }
+            base.OnHandleCreated(e);
+        }
+
         protected override void WndProc(ref Message m)
         {
             //Notify the dockable panel was closed